feat. Create Player Charactor
Player 캐릭터 선정

+++ Assets/Photon/PhotonRealtime/Code/LoadbalancingPeer.cs
@@ -1831,7 +1831,7 @@ namespace Photon.Realtime
         private bool isOpen = true;
 
         /// <summary>Max number of players that can be in the room at any time. 0 means "no limit".</summary>
-        public int MaxPlayers;
+        public byte MaxPlayers;
 
         /// <summary>Time To Live (TTL) for an 'actor' in a room. If a client disconnects, this actor is inactive first and removed after this timeout. In milliseconds.</summary>
         public int PlayerTtl;
